Add a post-build-hook
Add a new `post-build-hook` mechanism such that passing `--option
post-build-hook /foo/bar` to a nix-* command will cause `/foo/bar
/nix/store/drv-that/has-been-built.drv -- out1 out2 ... outn` to be
executed after each build.

One specially interesting use-case for this is to upload all the builded
artifacts to a binary cache (including the ones that don't appear in the
runtime closure of the final derivation or are built because of IFD).